Add description and update documentation for 'path' positional argument
The 'git-aicommit' command has been updated to provide a more descriptive explanation for the 'path' positional argument. This change modifies the README.md file to include an example usage of the command and updates the documentation for the 'path' option in the main function. The new description explains that this option is used to specify the directory path for committing.